--- a/src/mem.h
+++ b/src/mem.h
@@ -30,6 +30,8 @@ struct cpu_t;
 typedef uint8_t (*mem_mod_t)(struct cpu_t *cpu, uint8_t b);
 
 uint8_t mem_get_byte(struct cpu_t *cpu, uint16_t addr);
+void mem_set_byte(struct cpu_t *cpu, uint16_t addr, uint8_t v);
+
 uint8_t mem_get_byte_abs(struct cpu_t *cpu, uint16_t addr);
 uint8_t mem_get_byte_absx(struct cpu_t *cpu, uint16_t addr);
 uint8_t mem_get_byte_absy(struct cpu_t *cpu, uint16_t addr);
@@ -40,7 +42,6 @@ uint8_t mem_get_byte_indx(struct cpu_t *cpu, uint8_t addr);
 uint8_t mem_get_byte_indy(struct cpu_t *cpu, uint8_t addr);
 uint8_t mem_get_byte_ind(struct cpu_t *cpu, uint8_t addr);
 
-void mem_set_byte(struct cpu_t *cpu, uint16_t addr, uint8_t v);
 void mem_set_byte_zpg(struct cpu_t *cpu, uint8_t addr, uint8_t v);
 void mem_set_byte_zpgx(struct cpu_t *cpu, uint8_t addr, uint8_t v);
 void mem_set_byte_zpgy(struct cpu_t *cpu, uint8_t addr, uint8_t v);