## Raspberry Pi Zero

### Before

```
TEST Running 6502 tests
TEST   Success; executed 88957316 cycles in 6.4370 at 13.8197 MHz
TEST Running 65C02 tests
TEST   Success; executed 65094140 cycles in 4.6070 at 14.1294 MHz
```

### After

```
TEST Running 6502 tests
TEST   Success; executed 88957316 cycles in 3.6460 at 24.3986 MHz
TEST Running 65C02 tests
TEST   Success; executed 65094140 cycles in 2.5300 at 25.7289 MHz
```

## MacBook Pro 13"

### Before

```
TEST Running 6502 tests
TEST   Success; executed 88957316 cycles in 0.6490 at 137.0683 MHz
TEST Running 65C02 tests
TEST   Success; executed 65094140 cycles in 0.4100 at 158.7662 MHz
```

### After

```
TEST Running 6502 tests
TEST   Success; executed 88957316 cycles in 0.4470 at 199.0097 MHz
TEST Running 65C02 tests
TEST   Success; executed 65094140 cycles in 0.2930 at 222.1643 MHz
```